backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "anthropic>=0.50.0",
    "colorlog>=6.9.0",
    "docling>=2.58.0",
    "easyocr>=1.7.2",
    "fastapi>=0.116.1",
    "instructor>=1.7.9",
    "jinja2>=3.1.6",
    "openai>=1.77.0",
    "pillow>=11.2.1",
    "pydantic>=2.11.7",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "python-frontmatter>=1.1.0",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.1",
    "sentence-transformers>=5.1.0",
    "uvicorn>=0.35.0",
]

[dependency-groups]
dev = [
    "httpx>=0.28.1",
    "pytest>=8.0.0",
]

[tool.pytest.ini_options]
pythonpath = "."
